A project has a single sound effects file
I originally thought it might be a good idea for a project to have
multiple sound effect files, but after thinking it over and planning the
GUI I have decided against it.

Multiple sound effect files would complicate the GUI.  I can see myself
accidentally adding a new sound effect to the wrong file and getting
frustrated with myself for doing so.

A single sound effect file will simplify the GUI and one of the goals of
this project is keep-it-simple.