Improve switch-to-loader IO protocol and add more switch-to-loader tests
Updated the switch-to-loader IO protocol so the same IO port write will
work in the loader and audio-driver.

Added a switch-to-loader test in the loader that will restart the loader
in the middle of a transfer.

Added a switch-to-loader test to the `end`, `end_loop` and
`return_from_subroutine` bytecode instructions to hopefully ensure the
loader-test is executed when the bytecode goes off the rails and never
sleeps.