Merge MML commands across multiple lines
To ensure the section tick count is correct the parser must track the
tick-count for every play-note, rest, portamento and broken-chord
command.

The parser does not update the tick-count after an EndLoop command as it
complicated and already handled by the Bytecode struct.  Instead the
MmlBytecodeGenerator will copy the `bc` tick-count whenever it
encounters an EndLoop command.