Fix missing set instrument/envelope bc instruction after a L MML command
In the following MML:

    @1 dummy_instrument
    @2 dummy_instrument adsr 10 2 2 10
    @3 different_instrument

    A @1 a L @2 b @3 c

The `@2` after the song's loop-point will erroneously emit a `set_adsr`
instruction (as `@1` and `@2` use the same instrument_id). The song ends
with different instrument_id to the `L` loop-point, causing the `b` note
to be played with the wrong BRR sample when the song starts looping.

This bug also affects the following MML commands after a L set loop
point command:
 * `@` set instrument MML commands that do not change the instrument
 * `A` set adsr MML commands that do not change the envelope
 * `G` set gain MML commands that do not change the envelope

Reported by KungFuFurby