Fix audio glitch caused by ducked channel writing to the S-DSP too early
Setting `voiceChannelsDirty_music` bits when a sfx channel is disabled
can cause S-DSP writes to a voice channel that is in the middle of the
release envelope, creating an audio glitch.

To fix this glitch, `voiceChannelsDirty_music` is now set in
`_process_sfx_channels__inline()` and I only enable a music channel when
there is no sfx key-off event.