Change bytecode format to increase number of bytecode instructions
 * Opcodes 0..=63 are non play-note instructions
 * Opcodes 64..=255 are play-note instructions

Moving INSTRUCTIONS_WITH_ARGUMENTS and NO_ARGUMENT_INSTRUCTIONS to the
start of the opcode list saves a subtraction in process_next_bytecode().

This commit also changes the pitch table.  The pitchTable and
`instruments_pitchOffset` are now offset by
`FIRST_PLAY_NOTE_INSTRUCTION/2` to avoid a subtraction when playing
notes in the audio driver.