Fix portamento pitch changes delayed 1 tick
I noticed I was processing portamento before bytecode, which caused the
portamento pitch slide to be delayed by 1 ticks.

Moving the portamento effect to after the countdown timer and bytecode
was processed ensures portamento effects are processed on the same tick
as the portamento instruction.

Vibrato effects are still processed before bytecode as I want the
vibrato to change the pitch 1 tick *after* a play-note instruction
sets the pitch.  To clarify, I want the play-note instruction's pitch
audible for a single tick before it changed by the vibrato effect.