Fix unbalanced tremolo and panbrello
The offset_per_tick was calculated for the first quarter of the
triangle wave and did not account for the 2*quarter_wavelength_ticks
down slide.  Resulting in an imbalance where the effect overshoots
the expected maximum by 1 (ie, max = start + amplitude + 1).

Changing the starting sub-value to 0x7f (u8::MAX / 2) and updating the
offset_per_tick calculation ensures the min/max of the tremolo/panbrello
are equally spaced away from the starting volume/pan.

This commit also fixes the debug asserts in _tremolo_panbrello() and
adds slide-down debug asserts to confirm the new offset_per_tick
calculation.

This commit also increases the maximum panbrello amplitude to 64.