Fix noise disabled during the interrupted/stopped sfx release envelope
If a sfx using the noise generator is interrupted or stopped (using
STOP_SOUND_EFFECTS IO command), then the NON shadow is cleared before
the key-off release envelope has completed.

This commit removes the nonShadow_sfx bit-clear in
__play_sfx_on_channel() and disable_channel().  This will not interfere
with the new sound-effect as noise is disabled on the next play_note or
play_pitch instruction.