Fix EON and NON when channels G & H are ducked and unducked
When song channel H is playing noise and a sound effect ducts the
channel H, the NON bit is cleared during the key-off release.  The same
issue occurs when the sound effect ends and the voice channel is
unducked.

This issue is caused by 2 bugs:
 1. An off by one when masking the EON and NON registers.
 2. Not masking the sound effect EON and NON bits.

Testing was done manually, by recording and the output of tad-gui and
manually reviewing the key-off release envelope.