Add basic PMOD previewing to tad-gui
To preview PMOD when the user clicks on the "Play from cursor" button,
this commit modifies the rust bytecode interpreter to:
 * Set the VxPITCH virtual channel variables.
 * Key-on PMOD source if the next channel has PMOD on.

Be aware, this code does the bare minimum to set the VxPITCH variables.
It does not emulate vibrato, not does it emulate portamento.  If the
last instruction is a portamento instruction it will set VxPITCH to the
portamento target.

I believe emulating pitch-effects is too much complexity and CPU-time
for very little gain (as the preview will never be accurate since
bc_interpreter cannot emulate the envelope).