Track two noise frequencies for SFX and handle frequency restoration
Noise frequencies must be tracked separately between the music and SFX
instances. Music still has one noise frequency, but each SFX instance must have
a separate noise frequency to track, as they run independently of each other,
unlike the music. This also means that a system must be in place to set the
frequencies according to whether or not the SFX is running and the priority.
Currently the interruptible flags will do the priority part, as the raw IDs
themselves are currently not stored in variable RAM when they are playing (they
are stored to be tested for retriggering a SFX instance in the same voice,
though).