Zero volume on voices that use noise but don't have frequency priority
This required quite a few set of changes to implement this:
- All music voices that use noise, as well as all SFX voices, now have their
volShadowDirty bits set when updating the noise frequencies.
- A noise check was added immediately after writing to NON that zeroes out the
VxVOL DSP registers of all voices that don't have noise frequency priority
(either music or one of the two SFX instances). This avoids a noise volume
stacking effect due to only having one frequency to control.
- _update_vol_shadow now has a noise check at the beginning in order to enforce
zeroing out of the volume. The volume is zeroed out if the channel is using
noise but is not the one currently controlling the hardware noise frequency. It
is restored when it is either not using noise anymore or it gains control of the
hardware noise frequency.
- volShadowDirty now always sets voiceChannelsDirty_tmp after _update_vol_shadow
is processed. Without doing this, the volume updates were occurring later than
expected, particularly since the volume restoration relies on running
_update_vol_shadow then updating the DSP registers instead of backing them up
and restoring them.
- The SFX VxVOL mirrors are also zeroed out when the voice doesn't have noise
frequency priority. This is so that voiceChannelsDirty_tmp doesn't set the
volume to something non-zero by mistake, and cause noise stacking in the
process.

Additionally...
- Fixed a bug where the interruptibility flags were being handled backwards when
doing frequency and volume handling.
- Added an update call to the noise frequency setup when playing a note because
it disables noise.