Correct a bug where an uninitialized noise frequency can mute the sound
The initialization values for the noise frequencies has been changed to avoid a
bug where an uninitialized music noise frequency can result in a SFX instance
causing the entire sound to mute upon an attempted restoration of the music
noise frequency. Additionally, when preparing a noise frequency for restoration,
the value is ANDed by dsp.FLG__NOISE_FREQ_MASK instead of just simply clearing
the MSB as an extra safety measure.