Fix noise frequency updated before SFX VxVOL is zeroed
If a sound effect encounters a play_noise instruction and the other
sound effect was playing noise, new play_noise frequency would be
written to the S-DSP before the old noise was silenced with a VxVOL zero
write.

This can create a slightly auible glitch for a few samples, especially
if the two sound effects are played on opposite speakers.