Prefix 65816 API private variables with TadPrivate_
and tadPrivate_ for PvSnesLib (to match the tad_ public variables).

This ensures the private variables will have the same name for the 3
APIs.

I originally wanted to use a `_tad` prefix, but WLA-DX does not allow
.RAMSECTION variables starting with an underscore (they are local
symbols and cannot be used outside of the .SECTION or .RAMSECTION they
are defined in).