Fix music/sfx playing for a single tick after a pause IO command
A race condition exists where the IO command will stop the timer while
the counter is greater than zero.

This can cause a single music and/or sound effect tick to play after the
pause command.

If the next tick keys-on a channel with a fixed or incrementing GAIN
envelope, the audio driver will erroniously play a constant sound when
it should have been paused.