Rewrite unknown velocity portamento with 2 new bytecode instructions
This commit fixes `q` early-release and `Q` quantization with a
portamento with an unknown velocity.

`q` early release requires a slurred note-rest chain to end with
`rest 257`.  A unknown velocity portamento with a tie does not honour
this (`{cd} ^` outputs `portamento d4 no_keyoff 0 23 | rest 24`),
resulting in early-release activating at different times for portamentos
with a known and unknown velocity.

For `Q` quantization and an unknown velocity portamento - the audio
driver used the duration of the portamento instruction to calculate the
portamento velocity, which is different to the pitch slide length used
by the compiler's portamento speed calculation.  This causes the unknown
velocity portamento to reach the target VxPITCH when it should have been
cut mid pitch slide.

Both of these issues are solved by creating a `portamento_calc` and
`portamento_pitch_calc` bytecode instructions that includes the pitch
slide length (in ticks) as a parameter.