Disable portamento and MP pitch calculations if a song uses driver transpose
Determining if a portamento or MP vibrato is played with an unknown
transpose requires a rewrite of the MML compiler.  To expedite testing
of audio driver transpose I am temporarily disable pitch offset per tick
calculations if there is a single `set_transpose` or `adjust_transpose`
in the song.

The next production release of the audio driver will include an analysis
phase that will detect if a portamento or MP vibrato is played with an
unknown instrument and/or transpose value.