Disable note-range tests for bytecode assembly in the MML tests
This commit adds a feature flag that adds `UnitTestAssembly` and
`UnitTestAssemblySubroutine` BytecodeContext values that disable the
note-range and max-edl tests in the bytecode module.

These new contexts are required for the unit tests to pass after I move
the LoopAnalysis processing to the bytecode module.  LoopAnalysis
requires an analysed command stream and `BytecodeAssembler` does not
convert assembly to commands nor analyses them.  I also do not want the
unit tests to analyse or mutate the bytecode assembly in the tests.

I locked these new contexts behind a new "test" feature flag to prevent
it from impacting the performance of tad-compiler or tad-gui.

I also locked the `BytecodeAssembler` behind the "test" feature flag to
denote that it is intended for only the unit tests.