Add custom dialogs for the unsaved changes dialogs
The included fltk dialogs cannot be placed in the centre of the main
window.

Coding a custom modal window ensures the unsaved changes dialog box
opens in the middle of the screen.  This ensures the use has to move the
mouse to the center of the window when closing an unsaved tab.

The quit with unsaved files dialog now displays a list of all unsaved
files.